# routers/api/v1/article.go
package v1

import (
	"net/http"

	"github.com/astaxie/beego/validation"
	"github.com/gin-gonic/gin"
	"github.com/huanggengzhong/go-gin-example/models"
	"github.com/huanggengzhong/go-gin-example/pkg/e"
	"github.com/huanggengzhong/go-gin-example/pkg/logging"
	"github.com/huanggengzhong/go-gin-example/pkg/setting"
	"github.com/huanggengzhong/go-gin-example/pkg/util"
	"github.com/unknwon/com"
)

// ShowAccount godoc
//
//	@Summary		详情
//	@Description	文章详情
//	@Tags			文章
//	@Accept			json
//	@Produce		json
//	@Success		200 {string} json "{"code":200,"data":{},"msg":"ok"}"
//	@Router			/api/v1/articles/:id [get]
func GetArticle(c *gin.Context) {
	id := com.StrTo(c.Param("id")).MustInt()

	valid := validation.Validation{}
	valid.Min(id, 1, "id").Message("id必须大于0")

	code := e.INVALID_PARAMS
	var data interface{}
	if !valid.HasErrors() {
		if models.ExistArticleByID(id) {
			data = models.GetArticle(id)
			code = e.SUCCESS
		} else {
			code = e.ERROR_NOT_EXIST_ARTICLE
		}
	} else {
		for _, err := range valid.Errors {
			// log.Printf("err.Key:%s,err.Message:%s", err.Key, err.Message)
			logging.Info(err.Key, err.Message)
		}
	}
	c.JSON(http.StatusOK, gin.H{
		"code": code,
		"msg":  e.GetMsg(code),
		"data": data,
	})
}

// ShowAccount godoc
//
//	@Summary		列表
//	@Description	获取文章列表
//	@Tags			文章
//	@Accept			json
//	@Produce		json
//	@Param			tag_id	path		string	false	"标签id(tag_id必须大于0)"
//	@Param			state	path		int	false	"状态(只允许0或1) "
//	@Param			page	path		int	false	"起始页"
//	@Success		200 {string} json "{"code":200,"data":{},"msg":"ok"}"
//	@Router			/api/v1/articles [get]
func GetArticles(c *gin.Context) {
	code := e.INVALID_PARAMS
	data := make(map[string]interface{})
	maps := make(map[string]interface{})
	valid := validation.Validation{}
	var state int = -1
	if arg := c.Query("state"); arg != "" {
		state = com.StrTo(arg).MustInt()
		maps["state"] = state
		valid.Range(state, 0, 1, "state").Message("状态只允许0或1")
	}
	var tagId int = -1
	if arg := c.Query("tag_id"); arg != "" {
		tagId = com.StrTo(arg).MustInt()
		maps["tag_id"] = tagId
		valid.Min(tagId, 1, "tag_id").Message("tag_id必须大于0")
	}

	if !valid.HasErrors() {
		data["list"] = models.GetArticles(util.GetPage(c), setting.PageSize, maps)
		data["total"] = models.GetArticleTotal(maps)
		code = e.SUCCESS
	}
	c.JSON(http.StatusOK, gin.H{
		"code": code,
		"msg":  e.GetMsg(code),
		"data": data,
	})

}

// ShowAccount godoc
//
//	@Summary		新增
//	@Description	新增文章
//	@Tags			文章
//	@Accept			json
//	@Produce		json
//	@Param			title	path		string	true	"标题"
//	@Param			content	path		string	true	"内容"
//	@Param			desc	path		string	true	"简介"
//	@Param			created_by	path		string	true	"创建人"
//	@Param			tag_id	path		string	true	"标签id(tag_id必须大于0)"
//	@Param			state	path		int	true	"状态(只允许0或1) "
//	@Success		200 {string} json "{"code":200,"data":{},"msg":"ok"}"
//	@Router			/api/v1/articles [post]
func AddArticle(c *gin.Context) {
	tagId := com.StrTo(c.Query("tag_id")).MustInt()
	title := c.Query("title")
	desc := c.Query("desc")
	content := c.Query("content")
	createBy := c.Query("created_by")
	state := com.StrTo(c.DefaultQuery("state", "0")).MustInt()

	valid := validation.Validation{}
	valid.Min(tagId, 1, "tag_id").Message("tag_id必须大于0")
	valid.Required(title, "title").Message("标题不能为空")
	valid.Required(desc, "desc").Message("简介不能为空")
	valid.Required(content, "content").Message("内容不能为空")
	valid.Required(createBy, "created_by").Message("创建人不能为空")
	valid.Range(state, 0, 1, "state").Message("状态只允许0或1")

	code := e.INVALID_PARAMS
	if !valid.HasErrors() {

		if models.ExistTagByID(tagId) {
			data := make(map[string]interface{})
			data["tag_id"] = tagId
			data["title"] = title
			data["desc"] = desc
			data["content"] = content
			data["created_by"] = createBy
			data["state"] = state
			models.AddArticle(data)
			code = e.SUCCESS
		} else {
			code = e.ERROR_NOT_EXIST_TAG
		}
	} else {
		for _, err := range valid.Errors {
			// log.Printf("err.key是:%s,err.message是:%s", err.Key, err.Message)
			logging.Info(err.Key, err.Message)
		}
	}
	c.JSON(http.StatusOK, gin.H{
		"code": code,
		"msg":  e.GetMsg(code),
		"data": make(map[string]interface{}),
	})

}

// ShowAccount godoc
//
//	@Summary		编辑
//	@Description	编辑文章
//	@Tags			文章
//	@Accept			json
//	@Produce		json
//	@Param			tag_id	path		string	true	"标签id(tag_id必须大于0)"
//	@Param			title	path		string	true	"标题(最长为100字符)"
//	@Param			content	path		string	true	"内容(最长为65535字符)"
//	@Param			desc	path		string	true	"简介"
//	@Param			modified_by	path		string	true	"修改人(最长为100字符)"
//	@Param			state	path		int	true	"状态(只允许0或1) "
//	@Success		200 {string} json "{"code":200,"data":{},"msg":"ok"}"
//	@Router			/api/v1/articles/:id [put]
func EditArticle(c *gin.Context) {
	id := com.StrTo(c.Param("id")).MustInt()
	tagId := com.StrTo(c.Query("tag_id")).MustInt()
	title := c.Query("title")
	desc := c.Query("desc")
	content := c.Query("content")
	modifiedBy := c.Query("modified_by")

	var state int = -1
	valid := validation.Validation{}
	if arg := c.Query("state"); arg != "" {
		state = com.StrTo(arg).MustInt()
		valid.Range(state, 0, 1, "state").Message("状态只允许0或1")
	}
	valid.Min(id, 1, "id").Message("id必须大于0")
	valid.Min(tagId, 1, "tag_id").Message("tag_id必须大于0")
	valid.MaxSize(title, 100, "title").Message("标题最长为100字符")
	valid.MaxSize(desc, 255, "desc").Message("简介最长为255字符")
	valid.MaxSize(content, 65535, "content").Message("内容最长为65535字符")
	valid.Required(modifiedBy, "modified_by").Message("修改人不能为空")
	valid.MaxSize(modifiedBy, 100, "modified_by").Message("修改人最长为100字符")

	code := e.INVALID_PARAMS
	if !valid.HasErrors() {
		if models.ExistArticleByID(id) {

			if models.ExistTagByID(tagId) {
				data := make(map[string]interface{})
				if tagId > 0 {

					data["tag_id"] = tagId
				}
				if title != "" {
					data["title"] = title
				}
				if desc != "" {

					data["desc"] = desc
				}
				if content != "" {

					data["content"] = content
				}
				data["modified_by"] = modifiedBy
				models.EditArticle(id, data)
				code = e.SUCCESS
			} else {
				code = e.ERROR_NOT_EXIST_TAG
			}
		} else {
			code = e.ERROR_NOT_EXIST_ARTICLE
		}
	} else {
		for _, err := range valid.Errors {
			// log.Printf("err.Key:%s,err.Message:%s", err.Key, err.Message)
			logging.Info(err.Key, err.Message)
		}
	}
	c.JSON(http.StatusOK, gin.H{
		"code": code,
		"msg":  e.GetMsg(code),
		"data": make(map[string]interface{}),
	})

}

// ShowAccount godoc
//
//	@Summary		删除
//	@Description	删除文章
//	@Tags			文章
//	@Accept			json
//	@Produce		json
//	@Success		200 {string} json "{"code":200,"data":{},"msg":"ok"}"
//	@Router			/api/v1/articles/:id [delete]
func DeleteArticle(c *gin.Context) {
	id := com.StrTo(c.Param("id")).MustInt()
	valid := validation.Validation{}

	valid.Min(id, 1, "id").Message("id必须大于0")
	code := e.INVALID_PARAMS
	if !valid.HasErrors() {
		if models.ExistArticleByID(id) {

			models.DeleteArticle(id)
			code = e.SUCCESS
		} else {
			code = e.ERROR_NOT_EXIST_ARTICLE
		}
	} else {
		for _, err := range valid.Errors {
			// log.Printf("err.Key:%s,err.Message:%s", err.Key, err.Message)
			logging.Info(err.Key, err.Message)
		}
	}
	c.JSON(http.StatusOK, gin.H{
		"code": code,
		"msg":  e.GetMsg(code),
		"data": make(map[string]interface{}),
	})
}